Analysis

ANOVAapprox.get_AttributeRanking — Method

get_AttributeRanking( a::approx, λ::Float64 )::Vector{Float64}

This function returns the attribute ranking of the approximation for reg. parameter $\lambda$ as a vector of length a.d.

ANOVAapprox.get_AttributeRanking — Method

get_AttributeRanking( a::approx, λ::Float64 )::Dict{Float64,Vector{Float64}}

This function returns the attribute ranking of the approximation for all reg. parameters $\lambda$ as a dictionary of vectors of length a.d.

ANOVAapprox.get_GSI — Method

get_GSI( a::approx, λ::Float64; dict::Bool = false )::Union{Vector{Float64},Dict{Vector{Int},Float64}}

This function returns the global sensitivity indices of the approximation with $\lambda$ as a vector for dict = false or else a dictionary.

ANOVAapprox.get_GSI — Method

get_GSI( a::approx; dict::Bool = false )::Dict{Float64,Union{Vector{Float64},Dict{Vector{Int},Float64}}}

This function returns the global sensitivity indices of the approximation for all $\lambda$ as a vector for dict = false or else a dictionary.

ANOVAapprox.get_ShapleyValues — Method

get_ShapleyValues( a::approx, λ::Float64 )::Vector{Float64}

This function returns the Shapley values of the approximation for reg. parameter $\lambda$ as a vector of length a.d.

ANOVAapprox.get_ShapleyValues — Method

This function returns the Shapley values of the approximation for all reg. parameters $\lambda$ as a dictionary of vectors of length a.d.

ANOVAapprox.get_variances — Method

get_variances( a::approx, λ::Float64; dict::Bool = false )::Union{Vector{Float64},Dict{Vector{Int},Float64}}

This function returns the variances of the approximated ANOVA terms with $\lambda$ as a vector for dict = false or else a dictionary.

ANOVAapprox.get_variances — Method

get_variances( a::approx; dict::Bool = false )::Dict{Float64,Union{Vector{Float64},Dict{Vector{Int},Float64}}}

This function returns the variances of the approximated ANOVA terms for all $\lambda$ as a vector for dict = false or else a dictionary.