Errors

ANOVAapprox.get_L2error — Method

get_L2error( a::approx, norm::Float64, bc_fun::Function, λ::Float64 )::Float64

This function computes the relative $L_2$ error of the function given the norm norm and a function that returns the basis coefficients bc_fun for regularization parameter λ.

ANOVAapprox.get_L2error — Method

get_L2error( a::approx, norm::Float64, bc_fun::Function )::Dict{Float64,Float64}

This function computes the relative $L_2$ error of the function given the norm norm and a function that returns the basis coefficients bc_fun for all regularization parameters.

ANOVAapprox.get_l2error — Method

get_l2error( a::approx, λ::Float64 )::Float64

This function computes the relative $\ell_2$ error on the training nodes for regularization parameter λ.

ANOVAapprox.get_l2error — Method

get_l2error( a::approx, X::Matrix{Float64}, y::Union{Vector{ComplexF64},Vector{Float64}}, λ::Float64 )::Float64

This function computes the relative $\ell_2$ error on the data X and y for regularization parameter λ.

ANOVAapprox.get_l2error — Method

get_l2error( a::approx, X::Matrix{Float64}, y::Union{Vector{ComplexF64},Vector{Float64}}, )::Dict{Float64,Float64}

This function computes the relative $\ell_2$ error on the data X and y for all regularization parameters.

ANOVAapprox.get_l2error — Method

get_l2error( a::approx )::Dict{Float64,Float64}

This function computes the relative $\ell_2$ error on the training nodes for all regularization parameters.

ANOVAapprox.get_mad — Method

get_mad( a::approx, λ::Float64 )::Float64

This function computes the mean absolute deviation (mad) on the training nodes for regularization parameter λ.

ANOVAapprox.get_mad — Method

get_mad( a::approx, X::Matrix{Float64}, y::Union{Vector{ComplexF64},Vector{Float64}}, λ::Float64 )::Float64

This function computes the mean absolute deviation (mad) on the data X and y for regularization parameter λ.

ANOVAapprox.get_mad — Method

get_mse( a::approx, X::Matrix{Float64}, y::Union{Vector{ComplexF64},Vector{Float64}}, )::Dict{Float64,Float64}

This function computes the mean absolute deviation (mad) on the data X and y for all regularization parameters.

ANOVAapprox.get_mad — Method

get_mad( a::approx )::Dict{Float64,Float64}

This function computes the mean absolute deviation (mad) on the training nodes for all regularization parameters.

ANOVAapprox.get_mse — Method

get_mse( a::approx, λ::Float64 )::Float64

This function computes the mean square error (mse) on the training nodes for regularization parameter λ.

ANOVAapprox.get_mse — Method

get_mse( a::approx, X::Matrix{Float64}, y::Union{Vector{ComplexF64},Vector{Float64}}, λ::Float64 )::Float64

This function computes the mean square error (mse) on the data X and y for regularization parameter λ.

ANOVAapprox.get_mse — Method

get_mse( a::approx, X::Matrix{Float64}, y::Union{Vector{ComplexF64},Vector{Float64}}, )::Dict{Float64,Float64}

This function computes the mean square error (mse) on the data X and y for all regularization parameters.

ANOVAapprox.get_mse — Method

get_mse( a::approx )::Dict{Float64,Float64}

This function computes the mean square error (mse) on the training nodes for all regularization parameters.